refactor: centralize matchesPattern utility
