+++ server/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "ES2020",
-    "module": "ES2020",
+    "module": "ES2022",
     "rootDir": "./",
     "outDir": "./dist",
     "moduleResolution": "node",
