+++ server/utils/patterns.ts
@@ -1,5 +1,1 @@
-export function matchesPattern(value: string, pattern: string): boolean {
-  const escapeRegex = (s: string) => s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  const regex = new RegExp('^' + pattern.split('*').map(escapeRegex).join('.*') + '$');
-  return regex.test(value);
-}
+export { matchesPattern } from '../shared/matchesPattern.js'
